problema resuelto en el aritemtico finalizar compresion

+++ trunk/Speaker/src/ar/com/datos/grupo5/compresion/aritmetico/ParCharProb.java
@@ -31,7 +31,7 @@ public class ParCharProb implements Comparable<ParCharProb>{
 	}
 
 
-	public void setProbabilidad(float probabilidad) {
+	public void setProbabilidad(double probabilidad) {
 		this.probabilidad = probabilidad;
 	}
